+++ src/types/options/index.ts
@@ -6,3 +6,4 @@ export * from './LightingOptions';
 export * from './PathTracingOptions';
 export * from './RendererOptions';
 export * from './RenderingOptions';
+export * from './LoadingIndicatorOptions';